Restructure routing for separation among endpoints
