+++ config/secrets.js
@@ -1,0 +1,7 @@
+/*
+ * You generally want to .gitignore this file to prevent important credentials from being stored on your public repo.
+ */
+module.exports = {
+    token : "secret-llama",
+    mongo_connection : "mongodb://localhost/mp4"
+};
