+++ routes/index.js
@@ -1,0 +1,7 @@
+/*
+ * Connect all of your endpoints together here.
+ */
+module.exports = function (app, router) {
+  app.use('/api', require('./home.js')(router));
+  app.use('/api', require('./llama.js')(router));
+};
